+++ Pod/Classes/BBForm/BBStyleView.h
@@ -15,8 +15,8 @@
 @interface BBStyleView : UIView
 
 @property (nonatomic) /*IBInspectable*/ UIColor *borderColor;
-@property (nonatomic) /*IBInspectable*/ NSInteger borderWidth;
-@property (nonatomic) /*IBInspectable*/ NSInteger cornerRadius;
+@property (nonatomic) /*IBInspectable*/ CGFloat borderWidth;
+@property (nonatomic) /*IBInspectable*/ CGFloat cornerRadius;
 
 -(void)setup;
